services/helx-ldap/requirements.txt
ldap3
ldif3
pyyaml
